requirements.txt
tensorflow==2.4.1
pandas==1.2.4
numpy==1.19.5
ccxt==1.50.51
matplotlib==3.4.2
pyyaml==5.4.1
pytest==6.2.4